+++ example/src/App.tsx
@@ -442,6 +442,9 @@ const styles = StyleSheet.create({
     backgroundColor: '#000000',
     position: 'absolute',
     width: '100%',
+    // Automatically sized to 50 on phones and 90 on tablets. When adaptiveBannerEnabled is on,
+    // sized to AppLovinMAX.getAdaptiveBannerHeightForWidth().
+    height: 'auto',
     bottom: Platform.select({
       ios: 36, // For bottom safe area
       android: 0,