+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+    * Depend on Android SDK 11.7.0 and iOS SDK 11.7.0.
     * Deprecated getConsentDialogState().
 ## 4.1.4
     * Fix NPE for accessing a null icon in the native ad on Android.